# backend/openapi_schema.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "DHCaaS API",
                 "description":  "\n    ## Data Health Check as a Service - Enterprise Backend\n    \n    **DHCaaS** provides comprehensive data quality monitoring, governance, \n    and compliance management for enterprise data teams.\n    \n    ### Key Features:\n    - Data Quality Scanning - Multi-dimensional quality analysis\n    - Compliance Monitoring - GDPR, CCPA, HIPAA compliance checks\n    - Predictive Analytics - ML-powered anomaly detection\n    - Quality Metrics - Completeness, validity, consistency, accuracy\n    - PII Detection - Automatic sensitive data identification\n    - PDF Reports - Professional compliance reports\n    - Scan Jobs Management - Asynchronous data source scanning\n    ",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/v1/incidents":  {
                                            "get":  {
                                                        "tags":  [
                                                                     "incidents",
                                                                     "Incidents"
                                                                 ],
                                                        "summary":  "List All Incidents",
                                                        "description":  "Retrieve paginated list of incidents with optional filtering by status and severity",
                                                        "operationId":  "list_incidents_api_v1_incidents_get",
                                                        "parameters":  [
                                                                           {
                                                                               "name":  "page",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  "@{type=integer; minimum=1; description=Page number (1-indexed); default=1; title=Page}",
                                                                               "description":  "Page number (1-indexed)"
                                                                           },
                                                                           {
                                                                               "name":  "page_size",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  "@{type=integer; maximum=100; minimum=1; description=Items per page; default=20; title=Page Size}",
                                                                               "description":  "Items per page"
                                                                           },
                                                                           {
                                                                               "name":  "status",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  "@{anyOf=System.Object[]; description=Filter by incident status; title=Status}",
                                                                               "description":  "Filter by incident status"
                                                                           },
                                                                           {
                                                                               "name":  "severity",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  "@{anyOf=System.Object[]; description=Filter by severity level; title=Severity}",
                                                                               "description":  "Filter by severity level"
                                                                           }
                                                                       ],
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Successful Response",
                                                                                      "content":  "@{application/json=}"
                                                                                  },
                                                                          "422":  {
                                                                                      "description":  "Validation Error",
                                                                                      "content":  "@{application/json=}"
                                                                                  }
                                                                      }
                                                    }
                                        },
                  "/api/v1/incidents/{incident_id}/acknowledge":  {
                                                                      "post":  {
                                                                                   "tags":  [
                                                                                                "incidents",
                                                                                                "Incidents"
                                                                                            ],
                                                                                   "summary":  "Acknowledge Incident",
                                                                                   "description":  "Mark an incident as acknowledged by a specific user",
                                                                                   "operationId":  "acknowledge_incident_api_v1_incidents__incident_id__acknowledge_post",
                                                                                   "parameters":  [
                                                                                                      {
                                                                                                          "name":  "incident_id",
                                                                                                          "in":  "path",
                                                                                                          "required":  true,
                                                                                                          "schema":  "@{type=string; title=Incident Id}"
                                                                                                      }
                                                                                                  ],
                                                                                   "requestBody":  {
                                                                                                       "required":  true,
                                                                                                       "content":  {
                                                                                                                       "application/json":  "@{schema=}"
                                                                                                                   }
                                                                                                   },
                                                                                   "responses":  {
                                                                                                     "200":  {
                                                                                                                 "description":  "Successful Response",
                                                                                                                 "content":  "@{application/json=}"
                                                                                                             },
                                                                                                     "422":  {
                                                                                                                 "description":  "Validation Error",
                                                                                                                 "content":  "@{application/json=}"
                                                                                                             }
                                                                                                 }
                                                                               }
                                                                  },
                  "/api/v1/incidents/{incident_id}/resolve":  {
                                                                  "post":  {
                                                                               "tags":  [
                                                                                            "incidents",
                                                                                            "Incidents"
                                                                                        ],
                                                                               "summary":  "Resolve Incident",
                                                                               "description":  "Mark an incident as resolved with resolution notes",
                                                                               "operationId":  "resolve_incident_api_v1_incidents__incident_id__resolve_post",
                                                                               "parameters":  [
                                                                                                  {
                                                                                                      "name":  "incident_id",
                                                                                                      "in":  "path",
                                                                                                      "required":  true,
                                                                                                      "schema":  "@{type=string; title=Incident Id}"
                                                                                                  }
                                                                                              ],
                                                                               "requestBody":  {
                                                                                                   "required":  true,
                                                                                                   "content":  {
                                                                                                                   "application/json":  "@{schema=}"
                                                                                                               }
                                                                                               },
                                                                               "responses":  {
                                                                                                 "200":  {
                                                                                                             "description":  "Successful Response",
                                                                                                             "content":  "@{application/json=}"
                                                                                                         },
                                                                                                 "422":  {
                                                                                                             "description":  "Validation Error",
                                                                                                             "content":  "@{application/json=}"
                                                                                                         }
                                                                                             }
                                                                           }
                                                              },
...
              },
...
}